requirements.txt
absl-py==2.4.0
accelerate==0.32.1
astunparse==1.6.3
bitsandbytes==0.49.2
cachetools==7.0.5
captum==0.8.0
certifi==2026.2.25
charset-normalizer==3.4.6
cloudpickle==3.1.2
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
cycler==0.12.1
Deprecated==1.3.1
einops==0.8.2
fairscale==0.4.13
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
ftfy==6.3.1
gast==0.7.0
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.78.0
h5py==3.14.0
huggingface-hub==0.17.3
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.5.0
kornia==0.8.2
kornia_rs==0.1.10
lazy-loader==0.5
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
neurhci==2.0.0
numba==0.64.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
open_clip_torch==3.3.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
protobuf==7.34.0
psutil==7.2.2
pycocotools==2.0.11
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
quantus==0.6.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentencepiece==0.2.1
shap==0.49.1
six==1.17.0
slicer==0.0.8
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.25
tokenizers==0.14.1
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.34.0
triton==3.6.0
ttach==0.0.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
wrapt==2.1.2
Xplique==1.3.3
